// src/visual-lab/types.ts
export interface LabParameters {
  planetPrimaryAmplitude: number;
  planetPrimaryFrequency: number;
  planetSecondaryAmplitude: number;
  planetSecondaryFrequency: number;
  planetSpeed: number;
  planetShellOpacity: number;
  planetShellTint: number;
  planetTransmission: number;
  planetCoreScale: number;
  planetCoreDarkness: number;
  planetCoreOpacity: number;
  planetCoreDeformation: number;
  planetEnergyStrength: number;
  planetEnergyRegionCount: number;
  planetEnergyScale: number;
  planetEnergySpeed: number;
  planetEnergyColorBalance: number;
  planetFresnelStrength: number;
  planetRimColorBalance: number;
  planetHaloIntensity: number;
  orbitRadius: number;
  orbitTiltX: number;
  orbitTiltY: number;
  orbitDepth: number;
  orbitCoreOpacity: number;
  orbitGlowOpacity: number;
  nearBrightness: number;
  farBrightness: number;
  orbitFadeVariation: number;
  orbitEnergyPointIntensity: number;
  orbitCoreWidth: number;
  orbitGlowWidth: number;
  orbitFadeSmoothness: number;
  centerScale: number;
  centerLobeFrequency: number;
  centerLobeAmplitude: number;
  centerSecondaryAmplitude: number;
  centerShellOpacity: number;
  centerCoreDarkness: number;
  centerEnergyStrength: number;
  centerRegionSeparation: number;
  centerCyanIntensity: number;
  centerVioletIntensity: number;
  centerShellThickness: number;
  centerHaloSize: number;
  centerHaloIntensity: number;
  fov: number;
  cameraDistance: number;
  exposure: number;
  bloom: number;
  backgroundHaze: number;
}

export const defaultLabParameters: LabParameters = {
  planetPrimaryAmplitude: 0.16,
  planetPrimaryFrequency: 0.92,
  planetSecondaryAmplitude: 0.025,
  planetSecondaryFrequency: 1.75,
  planetSpeed: 0.12,
  planetShellOpacity: 0.27,
  planetShellTint: 0.52,
  planetTransmission: 0.76,
  planetCoreScale: 0.61,
  planetCoreDarkness: 0.95,
  planetCoreOpacity: 0.94,
  planetCoreDeformation: 0.065,
  planetEnergyStrength: 0.9,
  planetEnergyRegionCount: 3,
  planetEnergyScale: 0.17,
  planetEnergySpeed: 0.16,
  planetEnergyColorBalance: 0.58,
  planetFresnelStrength: 0.76,
  planetRimColorBalance: 0.56,
  planetHaloIntensity: 0.18,
  orbitRadius: 1.42,
  orbitTiltX: 0.46,
  orbitTiltY: -0.3,
  orbitDepth: 0.62,
  orbitCoreOpacity: 0.34,
  orbitGlowOpacity: 0.035,
  nearBrightness: 0.82,
  farBrightness: 0.1,
  orbitFadeVariation: 0.3,
  orbitEnergyPointIntensity: 0.45,
  orbitCoreWidth: 0.0032,
  orbitGlowWidth: 0.011,
  orbitFadeSmoothness: 0.72,
  centerScale: 0.92,
  centerLobeFrequency: 0.78,
  centerLobeAmplitude: 0.2,
  centerSecondaryAmplitude: 0.035,
  centerShellOpacity: 0.3,
  centerCoreDarkness: 0.97,
  centerEnergyStrength: 0.96,
  centerRegionSeparation: 0.24,
  centerCyanIntensity: 0.92,
  centerVioletIntensity: 0.78,
  centerShellThickness: 0.18,
  centerHaloSize: 1.15,
  centerHaloIntensity: 0.24,
  fov: 35,
  cameraDistance: 6.4,
  exposure: 0.88,
  bloom: 0.22,
  backgroundHaze: 0.42,
};